rename package "intf" to "core"
