+++ pkg/core/devices.go
@@ -1,4 +1,4 @@
-package intf
+package core
 
 import "golang.zx2c4.com/wireguard/wgctrl/wgtypes"
