+++ pkg/core/kernel_linux.go
@@ -1,4 +1,4 @@
-package intf
+package core
 
 import (
 	"errors"
