+++ pkg/core/peer.go
@@ -1,4 +1,4 @@
-package intf
+package core
 
 import (
 	"context"
