+++ pkg/core/peer_handlers.go
@@ -1,4 +1,4 @@
-package intf
+package core
 
 import (
 	"fmt"
