+++ pkg/core/watch_linux.go
@@ -1,4 +1,4 @@
-package intf
+package core
 
 import (
 	"fmt"
